+++ src/core/hex.ts
