+++ src/core/hsb.ts
