+++ src/core/keywords.ts
