+++ src/core/rgb.ts
