--- src/utils/index.ts
@@ -1,8 +1,0 @@
-export * from './basic'
-export * from './types'
-export * from './keywords'
-export * from './rgb'
-export * from './hsl'
-export * from './hsb'
-export * from './hex'
-// export * from './utils'
